backend/requirements.txt
flask
flask-cors
numpy
tensorflow
pymongo
python-telegram-bot
twilio
requests
